Crate ll_stack

Structs

GenericStack

GenericStack<T> is a linked-list based implementation of a stack: It implements the trait Stack, i.e., methods Stack::push, Stack::pop, Stack::peek, and Stack::peek_mut.

IntoIter

Iter

IterMut

Traits

Iterators

We define trait Iterators to define a three iterators for GenericStack:

Stack

trait Stack<T> defines the method a generic monomorphic stack needs to implement. We require that that the values that we push to a stack implement a set of traits.